// src/common/model/AnimalBoneModel.ts
import {Point} from "@/common/interface/Point";

export interface AnimalBoneModel {
  mouse: Point;
  head: Point;
  cervical_spine: Point;
  left_shoulder: Point;
  left_elbow: Point;
  left_wrist: Point;
  left_finger: Point;
  right_shoulder: Point;
  right_elbow: Point;
  right_wrist: Point;
  right_finger: Point;
  pelvis: Point;
  left_waist: Point;
  left_knee: Point;
  left_heel: Point;
  left_toe: Point;
  right_waist: Point;
  right_knee: Point;
  right_heel: Point;
  right_toe: Point;
}